clang: use llvm recipe
Change this recipe to build just clang and clang-tools-extra, using the
LLVM provided by the llvm recipe.

This adds an 'extra-tools' PACKAGECONFIG (enabled by default) that
controls whether to build the clang-tools-extra project. This includes
clang-tidy and clangd, but the compile time and size for these
components is not insignificant.

Add a patch from upstream to support using native prebuilt tools (such
as clang-tblgen) when building standalone.

Add a patch that is being worked on with upstream to not rebuild clang-
tblgen if it has already been provided. This saves a little build time,
but more importantly for us resolves a static linking/uninative problem.

Remove the dependency on clang-cross in non-native builds by just
depending on llvm-native and using the un-prefixed tools directly.

(From OE-Core rev: ccd006957779ec3c7dd677d7fcd84394274e067a)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>