linux-firmware: Move all firmware into sub-packages
A lot of the firmware has already been broken out into sub-packages, but
a good number were still lingering in the base linux-firmware package.

This patch does two things:

1) Move all of the remaining firmware into sub-packages based on the
driver name in the WHENCE file or into existing sub-packages where a
file or two were missed.

2) Create a new REMOVE_UNLICENSED variable and do_install() logic to
allow for marking certain firmware files as something to remove and not
package up.  The intention is that in the WHENCE file, a number of
firmware files have no license information or very questionable
declarations as to what the license is for the firmware.

(From OE-Core rev: 4dc3cb02e918819eb5e77ef4fd64203677b2247e)

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>