fragments: add a 'root-login-without-password' fragment
Please see this for background/some discussion:
https://lists.openembedded.org/g/openembedded-architecture/topic/115913545

Care should be taken to not enable this by default, and especially not for
production images. Poky and oe-core default templates did it, and it was
not a good starting point. Hopefully the fragment name, and the description
that users will see when enabling the fragment will provide enough warning.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>