testimage.bbclass: check that root-login-without-password features are present
More or less all of testimage relies on console and ssh logins root without
password to work; previously this was enabled by default in poky and core, but
now that it isn't, testimage will error out with timeouts in both console and
ssh login attempts. This commit performs an earlier check and provides a hint to the
users about what they should do.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>